+++ cart/views.py
@@ -58,138 +58,3 @@ class CartView(LoginRequiredMixin, TemplateView):
                 "debug_mode": settings.MITOL_UE_PAYMENT_INTERSTITIAL_DEBUG,
             },
         )
-
-
-@method_decorator(csrf_exempt, name="dispatch")
-class CheckoutCallbackView(View):
-    """
-    Handles the redirect from the payment gateway after the user has completed
-    checkout. This may not always happen as the redirect back to the app
-    occasionally fails. If it does, then the payment gateway should trigger
-    things via the backoffice webhook.
-    """
-
-    def post_checkout_redirect(self, order_state, request):
-        """
-        Redirect the user with a message depending on the provided state.
-
-        Args:
-            - order_state (str): the order state to consider
-            - order (Order): the order itself
-            - request (HttpRequest): the request
-
-        Returns: HttpResponse
-        """
-        if order_state == Order.STATE.CANCELED:
-            return redirect_with_user_message(
-                reverse("cart"), {"type": USER_MSG_TYPE_PAYMENT_CANCELLED}
-            )
-        elif order_state == Order.STATE.ERRORED:
-            return redirect_with_user_message(
-                reverse("cart"), {"type": USER_MSG_TYPE_PAYMENT_ERROR}
-            )
-        elif order_state == Order.STATE.DECLINED:
-            return redirect_with_user_message(
-                reverse("cart"), {"type": USER_MSG_TYPE_PAYMENT_DECLINED}
-            )
-        elif order_state == Order.STATE.FULFILLED:
-            return redirect_with_user_message(
-                reverse("cart"),
-                {
-                    "type": USER_MSG_TYPE_PAYMENT_ACCEPTED,
-                },
-            )
-        else:
-            if not PaymentGateway.validate_processor_response(
-                settings.ECOMMERCE_DEFAULT_PAYMENT_GATEWAY, request
-            ):
-                log.info("Could not validate payment response for order")
-            else:
-                processor_response = PaymentGateway.get_formatted_response(
-                    settings.ECOMMERCE_DEFAULT_PAYMENT_GATEWAY, request
-                )
-            log.error(
-                (
-                    "Checkout callback unknown error for transaction_id %s, state"
-                    " %s, reason_code %s, message %s, and ProcessorResponse %s"
-                ),
-                processor_response.transaction_id,
-                order_state,
-                processor_response.response_code,
-                processor_response.message,
-                processor_response,
-            )
-            return redirect_with_user_message(
-                reverse("cart"),
-                {"type": USER_MSG_TYPE_PAYMENT_ERROR_UNKNOWN},
-            )
-
-    def post(self, request):
-        """
-        Handle successfully completed transactions.
-
-        This does a handful of things:
-        1. Verifies the incoming payload, which should be signed by the
-        processor
-        2. Finds and fulfills the order in the system (which should also then
-        clear out the stored basket)
-        3. Perform any enrollments, account status changes, etc.
-        """
-
-        with transaction.atomic():
-            order = api.get_order_from_cybersource_payment_response(request)
-            if order is None:
-                return HttpResponse("Order not found")
-
-            # Only process the response if the database record in pending status
-            # If it is, then we can process the response as per usual.
-            # If it isn't, then we just need to redirect the user with the
-            # proper message.
-
-            if order.state == Order.STATE.PENDING:
-                processed_order_state = api.process_cybersource_payment_response(
-                    request,
-                    order,
-                    POST_SALE_SOURCE_REDIRECT,
-                )
-
-                return self.post_checkout_redirect(processed_order_state, request)
-            else:
-                return self.post_checkout_redirect(order.state, request)
-
-
-class CheckoutInterstitialView(LoginRequiredMixin, TemplateView):
-    """
-    Redirects the user to the payment gateway.
-
-    This is a simple page that just includes the checkout payload, renders a
-    form and then submits the form so the user gets thrown to the payment
-    gateway. They can then complete the payment process.
-    """
-
-    template_name = "checkout_interstitial.html"
-
-    def get(self, request):
-        """Render the checkout interstitial page."""
-        try:
-            checkout_payload = api.generate_checkout_payload(request)
-        except ObjectDoesNotExist:
-            return HttpResponse("No basket")
-        if (
-            "country_blocked" in checkout_payload
-            or "no_checkout" in checkout_payload
-            or "purchased_same_courserun" in checkout_payload
-            or "purchased_non_upgradeable_courserun" in checkout_payload
-            or "invalid_discounts" in checkout_payload
-        ):
-            return checkout_payload["response"]
-
-        return render(
-            request,
-            self.template_name,
-            {
-                "checkout_payload": checkout_payload,
-                "form": checkout_payload["payload"],
-                "debug_mode": settings.MITOL_UE_PAYMENT_INTERSTITIAL_DEBUG,
-            },
-        )
